refactor: rename grid-sync to gridops
- Rename cmd/grid-sync directory to cmd/gridops
- Update all import paths and references
- Update Makefile build and install targets
- Update .goreleaser.yml binary configuration
- Update .gitignore to reflect new binary name
- Update command usage and help text in Cobra commands
- Update inline comments referencing grid-sync
