+++ .gitignore
@@ -1,5 +1,5 @@
 /fogg
-/grid-sync
+/gridops
 coverage.out
 dist
