+++ cmd/gridops/cmd/common.go
