+++ cmd/gridops/cmd/doctor.go
