+++ cmd/gridops/cmd/root.go
@@ -16,9 +16,9 @@ var opts options
 var excludeDirs []string
 
 var rootCmd = &cobra.Command{
-	Use:   "grid-sync",
+	Use:   "gridops",
 	Short: "Syncs Fogg markers with the Grid API",
-	Long: `grid-sync is a tool to synchronize local marker files (.grid-state.yaml)
+	Long: `gridops is a tool to synchronize local marker files (.grid-state.yaml)
 generated by Fogg with the remote Grid API.`,
 }
