+++ cmd/gridops/cmd/testdata/cdktf/cdk.tf.json
