+++ cmd/gridops/cmd/testdata/hcl1/main.tf
