+++ cmd/gridops/cmd/testdata/hcl2/fogg.tf
