+++ cmd/gridops/cmd/testdata/hcl2/main.tf
