+++ cmd/gridops/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/chanzuckerberg/fogg/cmd/grid-sync/cmd"
+	"github.com/chanzuckerberg/fogg/cmd/gridops/cmd"
 )
 
 func main() {
